% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{gu_content}
\alias{gu_content}
\title{Content}
\usage{
gu_content(query, ..., verbose = TRUE)
}
\arguments{
\item{query}{Search query string.}

\item{...}{Use to pass any other parameters to the API.}

\item{verbose}{}
}
\description{
See the
\href{https://open-platform.theguardian.com/documentation/search}{API docs} for
full details on the query options available for this API.
}
