% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardianapi-package.R
\docType{package}
\name{guardianapi}
\alias{guardianapi-package}
\title{Tge \code{guardianapi} package}
\description{
Access to 'The Guardian' open API
\url{https://open-platform.theguardian.com/}, containing all articles, video and
images published in the 'Guardian' from 1999 to the present. Users must
\href{https://open-platform.theguardian.com/access/}{register} and use an API key,
which can be saved with the \code{\link[=gu_api_key]{gu_api_key()}} function, or as the \code{GU_API_KEY}
environmental variable. Free users can make up
to 5,000 calls per day and 12 calls per second, and access all article text
and associated metadata.
Images and video require a commercial subscription.
}
